legal_advice_project/AI/requirements.txt
fastapi
uvicorn[standard]
google-cloud-aiplatform>=1.26.0
google-auth
requests
python-dotenv
google-cloud-firestore
httpx
